fix: spinner struct
